remove src & create array args
